+++ src/redux/middleware/success.ts
@@ -1,5 +1,5 @@
 import { Middleware, MiddlewareAPI } from '@reduxjs/toolkit';
-import { createSuccessToasterObject } from '../../utils/createToasterObject';
+import { createSuccessToast } from '../../utils/createToast';
 
 import { endpoints as userEndpoints } from '../api/userApi';
 import { addToast } from '../slices/settingsSlice';
@@ -8,14 +8,12 @@ export const successMiddleware: Middleware =
   (api: MiddlewareAPI) => (next) => (action) => {
     if (userEndpoints.updateUserProfile.matchFulfilled(action)) {
       api.dispatch(
-        addToast(createSuccessToasterObject('Profile updated successfully'))
+        addToast(createSuccessToast('Profile updated successfully'))
       );
     }
 
     if (userEndpoints.updateUserAvatar.matchFulfilled(action)) {
-      api.dispatch(
-        addToast(createSuccessToasterObject('Avatar updated successfully'))
-      );
+      api.dispatch(addToast(createSuccessToast('Avatar updated successfully')));
     }
 
     return next(action);
